Fixed bug of card action whether to save or delete.

+++ portifolio/static_url.py
@@ -48,7 +48,7 @@ def redirect_action_card(card, cardaction):
             return url_portfolio()
         else:
             return redirect(reverse('portifolio:home'))
-    else:
+    elif cardaction is False:
         if card.section == '1':
             card.delete()
             return url_about()
@@ -60,6 +60,17 @@ def redirect_action_card(card, cardaction):
             return url_services()
         elif card.section == '4':
             card.delete()
+            return url_portfolio()
+        else:
+            return redirect(reverse('portifolio:home'))
+    else:
+        if card.section == '1':
+            return url_about()
+        elif card.section == '2':
+            return url_curriculum()
+        elif card.section == '3':
+            return url_services()
+        elif card.section == '4':
             return url_portfolio()
         else:
             return redirect(reverse('portifolio:home'))
